+++ libraries/RobotOpen/ROEncoder.h
@@ -10,7 +10,7 @@ class ROEncoder
     ROEncoder(uint8_t);
 
     long read();
-    long readCPS();
+    float readCPS();
     void setSensitivity(uint16_t sensitivity);
     void reset();
